+++ src/processsignaloperation.hh
@@ -31,9 +31,9 @@ class FreeboxProcessSignalOperation :
 
 public:
     ~FreeboxProcessSignalOperation() {}
-    void killProcess(int pid);
+    void killProcess(qint64 pid);
     void killProcess(const QString &filePath);
-    void interruptProcess(int pid);
+    void interruptProcess(qint64 pid);
     void interruptProcess(const QString &filePath);
 };
